+++ src/index.js
@@ -4,15 +4,13 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-import Stepper from "./components/mdl/mdl-stepper";
-import Selectfield from "./components/mdl/mdl-selectfield";
-import ListItemContentEx from "./components/mdl/mdl-listItemContentEx";
-import ContentEditable from "./components/mdl/mdl-contentEditable";
-import ExpansionPanel from "./components/mdl/mdl-expansionPanel";
+import Stepper from "./components/stepper";
+import ListItemContentEx from "./components/listItemContentEx";
+import ContentEditable from "./components/contentEditable";
+import ExpansionPanel from "./components/expansionPanel";
 
 export {
   Stepper,
-  Selectfield,
   ListItemContentEx,
   ContentEditable,
   ExpansionPanel,
